+++ pandora2d/refinement_cpp/include/dichotomy.hpp
@@ -144,7 +144,7 @@ struct Cost_volume_size {
  * @param cv_size : the structure containing the dimensions of the cost volume
  * @return Eigen::MatrixXd of size nb_disp_row * nb_disp_col
  */
-Eigen::MatrixXd get_cost_surface(const Eigen::VectorXd& cost_volume,
+Eigen::MatrixXd get_cost_surface(const Eigen::VectorXf& cost_volume,
                                  unsigned int index,
                                  Cost_volume_size& cv_size);
 
@@ -183,7 +183,7 @@ void search_new_best_point(const Eigen::MatrixXd& cost_surface,
  * @param filter : interpolation filter
  * @param method_matching_cost : max or min
  */
-void compute_dichotomy(Eigen::VectorXd& cost_volume,
+void compute_dichotomy(Eigen::VectorXf& cost_volume,
                        Eigen::Ref<Eigen::VectorXd> disparity_map_col,
                        Eigen::Ref<Eigen::VectorXd> disparity_map_row,
                        Eigen::Ref<Eigen::VectorXd> score_map,
